+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Removed
+- `Python 3.7` support, by @HardNorth
 
 ## [5.5.10]
 ### Added